requirements.txt
streamlit==1.16.0
numpy==1.24.1
matplotlib==3.6.2
krakenex==2.1.0
pandas==1.5.2
streamlit-option-menu==0.3.2
plotly.express==5.11.10
pykrakenapi==0.3.1